# Speed up the products table publication