--- a/client/modules/i18n/currency.js
+++ b/client/modules/i18n/currency.js
@@ -1,6 +1,5 @@
-import accounting from "accounting-js";
-import { Reaction, Logger } from "/client/api";
-import ReactionError from "@reactioncommerce/reaction-error";
+import { formatMoney } from "accounting-js";
+import { Reaction } from "/client/api";
 import { Shops, Accounts } from "/lib/collections";
 import { currencyDep } from "./main";
 
@@ -48,17 +47,9 @@ export function findCurrency(defaultCurrency, useDefaultShopCurrency) {
  * @memberof i18n
  * @method
  * @param {String} formatPrice - currentPrice or "xx.xx - xx.xx" formatted String
- * @param {Boolean} useDefaultShopCurrency - flag for displaying shop's currency in Admin view of PDP
  * @return {String} returns locale formatted and exchange rate converted values
  */
-export function formatPriceString(formatPrice, useDefaultShopCurrency) {
-  let defaultShopCurrency = useDefaultShopCurrency;
-
-  // in case useDefaultShopCurrency is a Spacebars.kw we have this check
-  if (typeof useDefaultShopCurrency === "object" || !useDefaultShopCurrency) {
-    defaultShopCurrency = false;
-  }
-
+export function formatPriceString(formatPrice) {
   currencyDep.depend();
   const locale = Reaction.Locale.get();
 
@@ -72,93 +63,47 @@ export function formatPriceString(formatPrice, useDefaultShopCurrency) {
   }
 
   // get user currency instead of locale currency
-  const userCurrency = findCurrency(locale.currency, defaultShopCurrency);
+  const userCurrency = findCurrency(locale.currency, true);
 
-  // for the cases then we have only one price. It is a number.
   const currentPrice = formatPrice.toString();
-  let price = 0;
   const prices = currentPrice.indexOf(" - ") >= 0 ?
-    currentPrice.split(" - ") : [currentPrice];
+    currentPrice.split(" - ") : [currentPrice, currentPrice];
 
-  // basic "for" is faster then "for ...of" for arrays. We need more speed here
-  const len = prices.length;
-  for (let i = 0; i < len; i += 1) {
-    const originalPrice = prices[i];
-    try {
-      // we know the locale, but we don"t know exchange rate. In that case we
-      // should return to default shop currency
-      if (typeof userCurrency.rate !== "number") {
-        throw new ReactionError("invalid-exchange-rate", "Exchange rate is invalid");
-      }
-      // Only convert for non-admin view.
-      if (!defaultShopCurrency) {
-        prices[i] *= userCurrency.rate;
-      }
-
-      price = _formatPrice(
-        price, originalPrice, prices[i],
-        currentPrice, userCurrency, i, len
-      );
-    } catch (error) {
-      Logger.debug("currency error, fallback to shop currency");
-      price = _formatPrice(
-        price, originalPrice, prices[i],
-        currentPrice, locale.shopCurrency, i, len
-      );
-    }
-  }
-  return price;
+  return getDisplayPrice(Number(prices[0]), Number(prices[1]), userCurrency);
 }
 
 /**
- * _formatPrice
- * private function for formatting locale currency
- * @private
- * @param  {Number} price         price
- * @param  {Number} originalPrice originalPrice
- * @param  {Number} actualPrice   actualPrice
- * @param  {Number} currentPrice  currentPrice
- * @param  {Number} currency      currency
- * @param  {Number} pos           position
- * @param  {Number} len           length
- * @return {Number}               formatted price
+ * @name getDisplayPrice
+ * @method
+ * @summary Returns a price for front-end display in the given currency
+ * @param {Number} minPrice Minimum price
+ * @param {Number} maxPrice Maximum price
+ * @param {Object} currencyInfo Currency object from Reaction shop schema
+ * @returns {String} Display price with currency symbol(s)
  */
-function _formatPrice(
-  price, originalPrice, actualPrice, currentPrice, currency,
-  pos, len
-) {
-  // this checking for locale.shopCurrency mostly
-  if (typeof currency !== "object") {
-    return false;
-  }
+function getDisplayPrice(minPrice, maxPrice, currencyInfo = { symbol: "" }) {
+  let displayPrice;
 
-  let adjustedPrice = actualPrice;
-  let formattedPrice;
-
-  // Precision is mis-used in accounting js. Scale is the propery term for number
-  // of decimal places. Let's adjust it here so accounting.js does not break.
-  if (currency.scale !== undefined) {
-    currency.precision = currency.scale;
-  }
-
-  // If there are no decimal places, in the case of the Japanese Yen, we adjust it here.
-  if (currency.scale === 0) {
-    adjustedPrice = actualPrice * 100;
-  }
-
-  // @param {string} currency.where: If it presents - in situation then two
-  // prices in string, currency sign will be placed just outside the right price.
-  // For now it should be manually added to fixtures shop data.
-  if (typeof currency.where === "string" && currency.where === "right" &&
-    len > 1 && pos === 0) {
-    const modifiedCurrency = Object.assign({}, currency, {
-      symbol: ""
-    });
-    formattedPrice = accounting.formatMoney(adjustedPrice, modifiedCurrency);
+  if (minPrice === maxPrice) {
+    // Display 1 price (min = max)
+    displayPrice = formatMoney(minPrice, currencyInfo);
   } else {
-    // accounting api: http://openexchangerates.github.io/accounting.js/
-    formattedPrice = accounting.formatMoney(adjustedPrice, currency);
+    // Display range
+    let minFormatted;
+
+    // Account for currencies where only one currency symbol should be displayed. Ex: 680,18 - 1 359,68 руб.
+    if (currencyInfo.where === "right") {
+      const modifiedCurrencyInfo = Object.assign({}, currencyInfo, {
+        symbol: ""
+      });
+      minFormatted = formatMoney(minPrice, modifiedCurrencyInfo).trim();
+    } else {
+      minFormatted = formatMoney(minPrice, currencyInfo);
+    }
+
+    const maxFormatted = formatMoney(maxPrice, currencyInfo);
+    displayPrice = `${minFormatted} - ${maxFormatted}`;
   }
 
-  return price === 0 ? currentPrice.replace(originalPrice, formattedPrice) : price.replace(originalPrice, formattedPrice);
+  return displayPrice;
 }