--- a/lib/api/catalog.js
+++ b/lib/api/catalog.js
@@ -19,51 +19,8 @@ export default {
    * @return {Object} range, min, max
    */
   getProductPriceRange(productId) {
-    const product = Products.findOne(productId);
-    if (!product) {
-      return {
-        range: "0",
-        min: 0,
-        max: 0
-      };
-    }
-
-    const variants = this.getTopVariants(product._id);
-    // if we have variants we have a price range.
-    // this processing will default on the server
-    const visibileVariant = variants.filter((variant) => variant.isVisible === true);
-
-    if (visibileVariant.length > 0) {
-      const variantPrices = [];
-      variants.forEach((variant) => {
-        if (variant.isVisible === true) {
-          const range = this.getVariantPriceRange(variant._id);
-          if (typeof range === "string") {
-            const firstPrice = parseFloat(range.substr(0, range.indexOf(" ")));
-            const lastPrice = parseFloat(range.substr(range.lastIndexOf(" ") + 1));
-            variantPrices.push(firstPrice, lastPrice);
-          } else {
-            variantPrices.push(range);
-          }
-        } else {
-          variantPrices.push(0, 0);
-        }
-      });
-      const priceMin = _.min(variantPrices);
-      const priceMax = _.max(variantPrices);
-      let priceRange = `${priceMin.toFixed(2)} - ${priceMax.toFixed(2)}`;
-      // if we don't have a range
-      if (priceMin === priceMax) {
-        priceRange = priceMin.toFixed(2);
-      }
-      return {
-        range: priceRange,
-        min: priceMin,
-        max: priceMax
-      };
-    }
-
-    if (!product.price) {
+    const product = Products.findOne({ _id: productId });
+    if (!product || !product.price) {
       return {
         range: "0",
         min: 0,