--- a/lib/api/catalog.js
+++ b/lib/api/catalog.js
@@ -1,4 +1,3 @@
-import _ from "lodash";
 import { Products } from "/lib/collections";
 
 /**