Impact: **minor**  
Type: **performance**

## Issue
Loading and paging the products table in the operator UI is slow with large Products collection.

## Solution

Results testing Mongo queries directly with 64654 top-level products:

```
Skip 0, no sorting: ~1ms
Skip 24000, no sorting: ~300ms
Skip 0, sort by createdAt: ~15ms
Skip 24000, sort by createdAt: ~1100ms
```

Sorting slows things down a bit as expected, but it is using the index so I don’t think there’s much to be done there. And 15ms query on the first page is respectable.

The query (without any fancy filtering) seems fine.

When I comment out the `Counts.publish` the table is much faster. So I think most of the blame is with publishing the count. We could look into using https://atmospherejs.com/natestrauser/publish-performant-counts package instead. But since we're soon moving this to GraphQL, I'd rather not introduce a new Meteor package.

For now, I'm seeing improvements by adding `noReady` and `nonReactive` options to the counts publication, and removing variants from the result set. We no longer need the variants included due to recent UI changes.

### Other changes
With the variants no longer being published, this exposed an issue with the client side money formatting where there would end up being two currency symbols in the formatted string. I rewrote that function to use code similar to what we use on the server.

## Breaking changes
None really but any custom plugins with operator UI components that are relying on variants being published might need updating.

## Testing
Go to the Products table in operator UI with a large number of top-level products in the database. Verify that it loads in a few seconds and takes only a few seconds to switch between pages.